LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window-expected.txt


FAIL page-created Error (structuredClone()) The object can not be cloned.
FAIL page-created Error (worker) The object can not be cloned.
FAIL page-created Error (cross-site iframe) The object can not be cloned.
FAIL page-created Error (same-origin iframe) The object can not be cloned.
PASS page-created DOMException (structuredClone())
PASS page-created DOMException (worker)
PASS page-created DOMException (cross-site iframe)
PASS page-created DOMException (same-origin iframe)
FAIL JS-engine-created TypeError (structuredClone()) The object can not be cloned.
FAIL JS-engine-created TypeError (worker) The object can not be cloned.
FAIL JS-engine-created TypeError (cross-site iframe) The object can not be cloned.
FAIL JS-engine-created TypeError (same-origin iframe) The object can not be cloned.
FAIL web API-created TypeError (structuredClone()) The object can not be cloned.
FAIL web API-created TypeError (worker) The object can not be cloned.
FAIL web API-created TypeError (cross-site iframe) The object can not be cloned.
FAIL web API-created TypeError (same-origin iframe) The object can not be cloned.
FAIL web API-created DOMException (structuredClone()) assert_equals: expected (string) "createElement@[native code]\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:33:27\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:41:31\n@http://localhost:8800/resources/testharness.js:2590:30\ntest@http://localhost:8800/resources/testharness.js:628:34\nstackTests@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:40:7\nglobal code@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:31:11" but got (undefined) undefined
FAIL web API-created DOMException (worker) assert_equals: expected (string) "createElement@[native code]\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:33:27\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:53:31\n@http://localhost:8800/resources/testharness.js:2590:30\nasync_test@http://localhost:8800/resources/testharness.js:676:38\nstackTests@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:52:13\nglobal code@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:31:11" but got (undefined) undefined
FAIL web API-created DOMException (cross-site iframe) assert_equals: expected (string) "createElement@[native code]\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:33:27\niframeTest@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:72:31\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:99:15\n@http://localhost:8800/resources/testharness.js:2590:30\nasync_test@http://localhost:8800/resources/testharness.js:676:38\nstackTests@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:96:13\nglobal code@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:31:11" but got (undefined) undefined
FAIL web API-created DOMException (same-origin iframe) assert_equals: expected (string) "createElement@[native code]\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:33:27\niframeTest@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:72:31\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:104:15\n@http://localhost:8800/resources/testharness.js:2590:30\nasync_test@http://localhost:8800/resources/testharness.js:676:38\nstackTests@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:102:13\nglobal code@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:31:11" but got (undefined) undefined
